refactor: split page into views and hooks
